installNode only installs node
Before it installed httpd and cloned and deployed git-repos. Now it only install the specified NodeJS version.
